EditSelectedSalesFloorViewをUIScrollViewとUIPageViewControllerを適用し、画面遷移まで実装
