+++ RemeApp/Controller/EditSalesFloorMapView/DefaultSalesFloorMapViewController.swift
@@ -1,0 +1,108 @@
+//
+//  DefaultSalesFloorMapViewController.swift
+//  RemeApp
+//
+//  Created by 本川晴彦 on 2023/04/15.
+//
+
+import UIKit
+
+class DefaultSalesFloorMapViewController: UIViewController {
+
+    /// 売り場のボタン：StoryboardのA-1
+    @IBOutlet private weak var greenThreeButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのA-2
+    @IBOutlet private weak var blueThreeButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのA-3
+    @IBOutlet private weak var redThreeButton: UIButton!
+    /// 売り場のボタン：StoryboardのA-3をタップして売り場の買い物リストに画面遷移
+
+    /// 売り場のボタン：StoryboardのB-1
+    @IBOutlet private weak var greenFourButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-2
+    @IBOutlet private weak var greenTwoButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-3
+    @IBOutlet private weak var blueSevenButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-4
+    @IBOutlet private weak var blueFourButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-5
+    @IBOutlet private weak var blueTwoButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-6
+    @IBOutlet private weak var redFourButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのB-7
+    @IBOutlet private weak var redTwoButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-1
+    @IBOutlet private weak var greenFiveButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-2
+    @IBOutlet private weak var greenOneButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-3
+    @IBOutlet private weak var blueSixButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-4
+    @IBOutlet private weak var blueFiveButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-5
+    @IBOutlet private weak var blueOneButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-6
+    @IBOutlet private weak var redFiveButton: UIButton!
+
+    /// 売り場のボタン：StoryboardのC-7
+    @IBOutlet private weak var redOneButton: UIButton!
+
+    /// レジのラベル
+    @IBOutlet private weak var registerLabel: UILabel!
+    /// 左出入り口のラベル
+    @IBOutlet private weak var leftEntranceLabel: UILabel!
+    /// 右出入り口のラベル
+    @IBOutlet private weak var rightEntranceLabel: UILabel!
+
+    @IBOutlet weak var leftCartView: UIImageView!
+
+    @IBOutlet weak var rightCartView: UIImageView!
+
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        
+    }
+    
+    /// レジ、左出入り口、右出入り口のラベルに枠線を設定するメソッド
+    private func setBorderAllLabel() {
+        registerLabel.setBorder()
+        leftEntranceLabel.setBorder()
+        rightEntranceLabel.setBorder()
+    }
+
+    /// 各UIButtonに購入商品の有無によって装飾を設定するメソッド
+    /// - 各ボタンに売り場の名称を設定
+    /// - 各ボタンに売り場の色を設定
+    /// - 各ボタンに装飾の設定
+    private func updateButtonAppearance() {
+        /// ボタンの配列を順番に設定
+        let buttons = [redOneButton, redTwoButton, redThreeButton, redFourButton, redFiveButton,
+                       blueOneButton, blueTwoButton, blueThreeButton, blueFourButton, blueFiveButton,
+                       blueSixButton, blueSevenButton, greenOneButton, greenTwoButton, greenThreeButton,
+                       greenFourButton, greenFiveButton]
+        // for文でbuttonsに順番にアクセス
+        for (index, button) in buttons.enumerated() {
+            let salesFloor = SalesFloorType(rawValue: index)!
+            button?.setTitle(salesFloor.nameOfSalesFloor, for: .normal)
+            button?.backgroundColor = salesFloor.colorOfSalesFloor
+            button?.setAppearanceWithShadow()
+        }
+    }
+
+}
